+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- support URLs in file field during `open` command (#47)
+
 
 ## [2.3.4] - 2020-09-14